+++ src/GPUAllocation.cu
@@ -1,7 +1,7 @@
 #include "GPUAllocation.h"
 
 /** move particle array to GPU */
-void particle_move2gpu(struct particles* part, struct particles** part_gpu, int n_streams, cudaStream_t* stream, int* offset, int* np_stream)
+void particle_move2gpu(struct particles* part, struct particles** part_gpu, int n_streams, cudaStream_t* stream, long* offset, long* np_stream)
 {
     // Allocate memory on the GPU
     cudaMalloc(part_gpu, sizeof(particles)); 
@@ -73,7 +73,7 @@ void particle_move2gpu(struct particles* part, struct particles** part_gpu, int 
 }
 
 /** move particle array to CPU */
-void particle_move2cpu(struct particles* part_gpu, struct particles* part, , int n_streams, cudaStream_t* stream, int* offset, int* np_stream)
+void particle_move2cpu(struct particles* part_gpu, struct particles* part, int n_streams, cudaStream_t* stream, long* offset, long* np_stream)
 {   
     // Create temporary copy of host pointers
     FPpart* x_host = part->x;
